requirements.txt
anyascii==0.3.2
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.11.2
black==22.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
curlylint==0.13.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.9
django-countries==7.5.1
django-crispy-forms==2.1
django-csp==3.7
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
djhtml==1.4.13
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
mypy-extensions==1.0.0
mysqlclient==2.2.4
openpyxl==3.1.2
parsy==1.1.0
pathspec==0.12.1
pillow==10.2.0
pillow-heif==0.14.0
platformdirs==4.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
ruff==0.0.290
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wagtail==5.2.2
wagtail-font-awesome-svg==1.0.1
webencodings==0.5.1
whitenoise==6.6.0
Willow==1.6.3
